code refactorisation - improved unit tests

+++ reg-lib/cpu/_reg_maths.cpp
@@ -807,10 +807,10 @@ mat44 reg_mat44_mul(mat44 const* A, mat44 const* B)
     {
         for (int j = 0; j < 4; j++)
         {
-            R.m[i][j] = A->m[i][0] * B->m[0][j] +
-                A->m[i][1] * B->m[1][j] +
-                A->m[i][2] * B->m[2][j] +
-                A->m[i][3] * B->m[3][j];
+            R.m[i][j] = static_cast<float>(static_cast<double>(A->m[i][0]) * static_cast<double>(B->m[0][j]) +
+                                           static_cast<double>(A->m[i][1]) * static_cast<double>(B->m[1][j]) +
+                                           static_cast<double>(A->m[i][2]) * static_cast<double>(B->m[2][j]) +
+                                           static_cast<double>(A->m[i][3]) * static_cast<double>(B->m[3][j]));
         }
     }
     return R;
